Update sequences.py
Changes to time_to_cycles() method of Sequence class:
1) Changed the parameter "wait_time" to "wait_simple".
The reason is that now we have two different waiting functions in toolkit. One is wait_zero(...) which uses playZero(...) in the sequencer and the other one is wait_simple(...) which uses wait(...) in sequencer. If we use wait_zero(...), then we need the number of cycles in terms of sampling rate (wait_simple=False). If we use wait_simple(...) command, then we need the number of cycles in terms of sequencer rate (wait_simple=True).
2) Changed default value of wait_simple from True to False.
That is because wait_zero(...) is the new default waiting function in toolkit.
3) Expanded the existing docstring to give more details about the wait_simple parameter.